Split ListCommand into ListExtensionsCommand and ListToolsCommand
