+++ src/command/list/namespace.ts
@@ -1,0 +1,1 @@
+export const namespace = 'list';
